refactor: readLineAsync view 폴더로 이동
